# Conflicts:
#	app/Services/YouthManagementServices/YouthService.php

+++ app/Services/YouthManagementServices/YouthProfileService.php
@@ -648,7 +648,7 @@ class YouthProfileService
             ],
             "skills.*" => [
                 "required",
-                'numeric',
+                'integer',
                 "distinct",
                 "min:1"
             ],
